requirements.txt
langchain==0.1.16
langchain_openai==0.1.3
python-dotenv==1.0.1
fastapi==0.110.2
uvicorn==0.29.0
redis==5.0.4
langchainhub==0.1.15
scikit-learn==1.5.0
streamlit==1.36.0
streamlit-extras==0.4.3

- e